+++ src/components/Characters/CharacterCard.js
@@ -4,8 +4,6 @@ import FetchCharacter from '../Fetch/FetchCharacter';
 import FetchAbilities from '../Fetch/FetchAbilities';
 import CharacterLevelUpForm from './CharacterLevelUpForm';
 import Belly from '../../assets/Belly.png';
-import '../../assets/font-awesome.min.css';
-import '../../assets/font-awesome-rules.min.css';
 
 const CharacterCard = ({ match, data }) => {
 	// Fetch the character's id
@@ -25,14 +23,16 @@ const CharacterCard = ({ match, data }) => {
 	if(Object.keys(character.data).length > 0) {
 		const data = character.data;
 		const tier = [];
-		const unlocked = data._unlock < account.current_stage;
+		const unlocked = data.stage_unlock < account.current_stage;
+		const is_purchased = data.is_purchased ?? 0;
 		
-		// Tier 1: 4
-		// Tier 2: 9
-		// Tier 3: 16
-		// Tier 4: 25
-		// Tier 5: 36
-		const max_lvl = Math.pow((data.tier + 1), 2);
+		// Tier 1: 5
+		// Tier 2: 10
+		// Tier 3: 20
+		// Tier 4: 40
+		// Tier 5: 80
+		const max_lvl = 5 * Math.pow(2, (data.tier - 1));
+		const current_lvl = data.level ?? 1;
 		
 		const calcStat = (lvl, tier, val) => {
 			return Math.round(Math.pow(lvl, 2) / 5 * (tier / 5) + val);
@@ -42,10 +42,10 @@ const CharacterCard = ({ match, data }) => {
 			return Math.pow(lvl, 2) + val;
 		}
 		
-		let hp = calcStat(data.level, data.tier, data.hp);
-		let attack = calcStat(data.level, data.tier, data.attack);
-		let defense = calcStat(data.level, data.tier, data.defense);
-		let cost = calcCost(data.level, 10);
+		let hp = calcStat(current_lvl, data.tier, data.hp);
+		let attack = calcStat(current_lvl, data.tier, data.attack);
+		let defense = calcStat(current_lvl, data.tier, data.defense);
+		let cost = calcCost(current_lvl, 10);
 		
 		for(let i = 1; i <= data.tier; i++) {
 			tier[i] = <i key={i} className="fas fa-star"></i>;
@@ -56,29 +56,24 @@ const CharacterCard = ({ match, data }) => {
 						<div className="left-panel">
 							<div className="top">
 								<div className="level-wrap">
-									<svg>
-										<g>
-											<path d="M25 0 L25 0 32 20 50 20 36 32 42 50 25 38 8 50 14 32 0 20 18 20 25 0" stroke="#000" strokeWidth="2" />
-											<circle cx="25" cy="28" r="15" stroke="#000" strokeWidth="2" />
-										</g>
-									</svg>
-									<span className="level" title="Current level">{data.level}</span>
+									<i className="far fa-circle"></i>
+									<span className="level" title="Current level">{current_lvl}</span>
 								</div>
-								<span className={'unlock' + (unlocked ? ' is-unlocked' : '')} title={unlocked ? 'Character unlocked!' : 'Character unlocks after Stage ' + data._unlock}><i className={unlocked ? 'fas fa-unlock' : 'fas fa-lock'}></i></span>
+								<span className={'unlock' + (unlocked ? ' is-unlocked' : '')} title={unlocked ? 'Character unlocked!' : 'Character unlocks after Stage ' + data.stage_unlock}><i className={unlocked ? 'fas fa-unlock' : 'fas fa-lock'}></i></span>
 							</div>
-							{unlocked && (data.cost === 0 || (data.cost > 0 && data.is_purchased)) && (account.medallions > 0 && account.belly > cost) ? (data.level < max_lvl ? <CharacterLevelUpForm ch_id={data.id} cost={cost} /> : <p>Max level reached!</p>) : null}
+							{unlocked && (data.cost === 0 || (data.cost > 0 && is_purchased)) && (account.medallions > 0 && account.belly > cost) ? (current_lvl < max_lvl ? <CharacterLevelUpForm ch_id={data.id} cost={cost} /> : <p>Max level reached!</p>) : null}
 							<ul className="stats">
 								<li><strong title="Hitpoints">HP</strong> <span>{hp}</span></li>
 								<li><strong title="Base attack">ATK</strong> <span>{attack}</span></li>
 								<li><strong title="Base defense">DEF</strong> <span>{defense}</span></li>
 								<li><strong title="Purchase cost">COST</strong> <span>{data.cost > 0 ? <span><img src={Belly} title="Belly" alt="Belly" />{data.cost}</span> : 'Free'}</span></li>
-								<li>{unlocked && data.cost > 0 && !data.is_purchased ? <Link to={`/characters/purchase/${id}/`} className="button">Purchase</Link> : null}</li>
+								<li>{unlocked && data.cost > 0 && !is_purchased ? <Link to={`/characters/purchase/${id}/`} className="button">Purchase</Link> : null}</li>
 							</ul>
 							<h2>Abilities</h2>
 							{abilities.isLoading ? <p>Loading...</p> : (abilities.data.length > 0 ? <ul className="abilities">
 								{abilities.data.map(ab => {
-									return <li key={ab.id} className={data.level < ab.lvl_unlock ? 'locked' : ''} title={`${ab.name}${data.level < ab.lvl_unlock ? ' \u2014 LOCKED' : ''}
-  Attack: ${ab.attack} • Defense: ${ab.defense} • Heal: ${ab.heal}
+									return <li key={ab.id} className={current_lvl < ab.lvl_unlock ? 'locked' : ''} title={`${ab.name}${current_lvl < ab.lvl_unlock ? ' \u2014 LOCKED' : ''}
+  ${ab._class.toUpperCase()}: ${ab._value} \u2022 CD: ${ab.cooldown} ${ab.cooldown === 1 ? 'turn' : 'turns'}
   Unlocks at: Lv${ab.lvl_unlock}
 
 ${ab.description}`}>{ab.name}</li>