+++ src/components/Characters/Characters.js
@@ -30,14 +30,14 @@ const Characters = ({ match }) => {
 								<th>Defense</th>
 								<th>Tier</th>
 								<th>Level</th>
-								<th>Type</th>
+								<th>Class</th>
 								<th>Unlocked?</th>
 								<th>Purchased?</th>
 							</tr>
 						</thead>
 						<tbody>
 							{characters.data.map(ch => {
-								return <Character key={ch.id} id={ch.id} name={ch.name} hp={ch.hp} attack={ch.attack} defense={ch.defense} tier={ch.tier} level={ch.level} type={ch.type} unlock={ch.unlock} stage={account.data.current_stage} cost={ch.cost} is_purchased={ch.is_purchased} />
+								return <Character key={ch.id} id={ch.id} name={ch.name} hp={ch.hp} attack={ch.attack} defense={ch.defense} tier={ch.tier} level={ch.level ?? 1} _class={ch._class} unlock={ch.unlock} stage={account.data.current_stage} cost={ch.cost} is_purchased={ch.is_purchased ?? 0} />
 							})}
 						</tbody>
 					</table>