+++ src/components/Fetch/FetchCharacter.js
@@ -1,7 +1,7 @@
 import { useState, useEffect, useCallback, useRef } from 'react';
 
 const FetchCharacter = (id) => {
-	const hasFetched = useRef(false);
+	const has_fetched = useRef(false);
 	const [character, setCharacter] = useState({data: [], isLoading: false});
 	
 	const fetchData = useCallback(async () => {
@@ -14,18 +14,18 @@ const FetchCharacter = (id) => {
 				throw new Error('Something went wrong!');
 			}
 			
-			const dataIn = await response.json();
+			const data_in = await response.json();
 			
-			setCharacter({data: dataIn, isLoading: false});
+			setCharacter({data: data_in, isLoading: false});
 		} catch(err) {
 			console.log(err.message);
 		}
 	}, [character.data, id]);
 	
 	useEffect(() => {
-		if(!hasFetched.current && id > 0) {
+		if(!has_fetched.current && id > 0) {
 			fetchData();
-			hasFetched.current = true;
+			has_fetched.current = true;
 		}
 	}, [fetchData, id]);